@@ -1,13 +1,12 @@
-cmake_minimum_required(VERSION 3.5)
+cmake_minimum_required(VERSION 3.11)
 project(SnD VERSION 1.0)
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 
 add_subdirectory(external/Argz)
-add_subdirectory(external/wRATter)
-add_subdirectory(external/spdlog)
+add_subdirectory(external/range-v3)
 add_subdirectory(external/nlopt)
-add_subdirectory(external/boost)
+#add_subdirectory(external/googletest)
 
 list(APPEND CMAKE_PREFIX_PATH $ENV{ROOTSYS})
 #---Locate the ROOT package and defines a number of variables (e.g. ROOT_INCLUDE_DIRS)
@@ -23,11 +22,11 @@ endif()
 # ## NOT NEEDED FOR DICT
 include_directories(
         ${PROJECT_SOURCE_DIR}/include
-        ${PROJECT_SOURCE_DIR}/external/wRATter/include
+#        ${PROJECT_SOURCE_DIR}/external/wRATter/include
         ${PROJECT_SOURCE_DIR}/external/Argz/include
         ${PROJECT_SOURCE_DIR}/external/ProgressBar/include
         ${PROJECT_SOURCE_DIR}/external/ROOT/include
-        ${PROJECT_SOURCE_DIR}/external/spdlog/include
+        ${PROJECT_SOURCE_DIR}/external/range-v3/include
         ${ROOT_INCLUDE_DIRS}
 )
 
@@ -39,15 +38,15 @@ include_directories(
 add_library(SVD src/LinAlg/SVD.cc)
 
 add_library(Hit src/SnD/Hit.cc)
-target_link_libraries(Hit SVD ROOT::Physics ROOT::Hist)
+#target_link_libraries(Hit SVD ROOT::Hist)
 
 add_library(TReader src/Templates/TReader.cc)
 
 add_library(NTuple src/Readers/NTuple.cc)
-target_link_libraries(NTuple TReader Hit ROOT::RIO ROOT::Tree ROOT::TreePlayer Boost::range Boost::any)
-
-add_library(Rat src/Readers/Rat.cc)
-target_link_libraries(Rat Hit wRATter)
+target_link_libraries(NTuple TReader Hit ROOT::RIO ROOT::Tree ROOT::TreePlayer)
+#
+#add_library(Rat src/Readers/Rat.cc)
+#target_link_libraries(Rat Hit wRATter)
 
 
 #----------------------------------------------------------------------------
@@ -59,13 +58,19 @@ target_link_libraries(TApp Argz NTuple)
 add_executable(CreatePDF src/Apps/CreatePDF.cc src/Analyzers/MakePDF.cc)
 target_link_libraries(CreatePDF Argz NTuple)
 
-add_executable(Recon src/Apps/Recon.cc src/Analyzers/Recon.cc
-        src/Algo/VHits.cc
-        src/Algo/WOpt.cc)
-target_link_libraries(Recon Argz NTuple Boost::foreach Boost::optional nlopt)
+#add_executable(Recon src/Apps/Recon.cc src/Analyzers/Recon.cc
+#        src/Algo/VHits.cc
+#        src/Algo/WOpt.cc)
+#target_link_libraries(Recon Argz NTuple Boost::foreach Boost::optional nlopt)
+#
+#add_executable(Map src/Apps/Map.cc src/Analyzers/Map.cc
+#        src/Algo/VHits.cc
+#        src/Algo/WOpt.cc
+#        src/SnD/SpaceGrid.cc)
+#target_link_libraries(Map Argz NTuple nlopt)
 
-add_executable(Map src/Apps/Map.cc src/Analyzers/Map.cc
-        src/Algo/VHits.cc
-        src/Algo/WOpt.cc
-        src/SnD/SpaceGrid.cc)
-target_link_libraries(Map Argz NTuple nlopt)
+#----------------------------------------------------------------------------
+# TESTS
+#
+#add_executable(Vector3Tests test/Vector3Tests.cc)
+#target_link_libraries(Vector3Tests gtest_main)
