@@ -6,12 +6,13 @@
 #define SND_INCLUDE_TEMPLATES_TDATA_HH_
 
 #include "SnD/Hit.hh"
+#include "SnD/Vector3.hh"
 
 class TData {
  protected:
  public:
-  virtual TVector3 GetPosition() = 0;
-  virtual TVector3 GetDirection() = 0;
+  virtual Vector3 GetPosition() = 0;
+  virtual Vector3 GetDirection() = 0;
   virtual double GetEnergy() = 0;
   virtual double GetTime() = 0;
   virtual std::vector<Hit> GetVHits() = 0;
