@@ -9,7 +9,7 @@
 //
 #include "../Readers/NTuple.hh"
 
-volatile sig_atomic_t TReader::gSignalStatus = 0;
+// volatile sig_atomic_t TReader::gSignalStatus = 0;
 
 int main(int argc, char **argv) {
 
